+++ data/processing/osc.py
@@ -1,9 +1,8 @@
-import re
-from typing import Optional
 import pandas as pd
 from traitlets import Enum
 from data.constants.raw_data_constants import OngsDatasetCols
 from data.constants.segmentation_code import SegmentationCode
+from data.processing.data_parser import valid_cnpj
 
 
 def osc_dataset(dataset: pd.DataFrame) -> pd.DataFrame:
@@ -99,26 +98,3 @@ def osc_segmentation_codes(entry: pd.Series) -> str:
     code_names = [code.name.replace("_", " ") for code in enum_codes]
 
     return ", ".join(code_names)
-
-
-def valid_cnpj(raw_cnpj: str | int) -> Optional[str]:
-    """
-    Validates and formats a CNPJ number.
-    Args:
-        raw_cnpj (str | int): The raw CNPJ number, which may contain non-digit characters.
-
-    Returns:
-        Optional[str]: The formatted CNPJ number (XX.XXX.XXX/XXXX-XX) if valid, otherwise None.
-    """
-
-    raw_cnpj_str = str(raw_cnpj)
-    digits: str = re.sub(r'\D', '', raw_cnpj_str)
-
-    zleading_cnpj = digits.zfill(14)
-
-    if len(zleading_cnpj) == 14:
-        return re.sub(r'^(\d{2})(\d{3})(\d{3})(\d{4})(\d{2})$',
-                      r'\1.\2.\3/\4-\5',
-                      zleading_cnpj)
-    else:
-        return None
